refactor(cli): consolidate hosted systemd management
